+++ src/main/java/model/elements/OvalElement.java
@@ -1,13 +1,11 @@
 package model.elements;
 
-import model.GraphicElement;
-
 import java.awt.*;
 
 public class OvalElement extends GraphicElement {
     private final int x, y, diameter;
 
-    public OvalElement(int x, int y, int diameter, Color color, float strokeWidth) {
+    public OvalElement(int x, int y, int diameter, Color color, int strokeWidth) {
         super(color, strokeWidth);
         this.x = x;
         this.y = y;
